Servicio Ventas Finalizado

+++ ventas-api-spring-boot/src/main/java/com/ventas/repository/VentaRepository.java
@@ -1,10 +1,9 @@
 package com.ventas.repository;
 
-
-import com.Ventas.models.Venta;
+import com.ventas.models.Venta;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VentaRepository extends JpaRepository<Venta, Integer> {
-}
+public interface VentaRepository extends JpaRepository<Venta, Long> {}
+
